move more docs to book, and simplify parachain-example
